+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Added validation for consumed notes when importing (#449).
 * Added the Web Client Crate
+* Added validations in transaction requests (#447).
 * [BREAKING] Refactored `Client` to merge submit_transaction and prove_transaction (#445)
 * Tracked token symbols with config file (#441).
 * [BREAKING] Refactored `TransactionRequest` to represent a generalized transaction (#438).